Move RegisterForm components to forms directory
Renamed and relocated RegisterForm.tsx and RegisterForm.module.scss from components/ to components/forms/ for better organization. Updated import paths in App.tsx and within the SCSS and TSX files to reflect the new structure.
