+++ printlib-frontend/src/components/forms/RegisterForm.module.scss
@@ -1,5 +1,5 @@
-@import '../styles/variables';
-@import './forms/LoginForm.module.scss';
+@import '../../styles/variables';
+@import './LoginForm.module.scss';
 
 .buttonRow {
     display: flex;
